Add files via upload
Welcome to my Back-end perspective web application, In this project I focused on back-end coding only it supports the Code First approach, Identity and Role management; by creating accounts so the Admin can add, view, delete  Role names, and view Users who are in this role. The Users are not authorized to do those operations!